+++ app/src/main/java/com/example/plantee/domain/model/Diagnosis.kt
@@ -3,12 +3,12 @@ package com.example.plantee.domain.model
 import java.time.LocalDateTime
 
 data class Diagnosis(
-    val id: Int = 0,
+    val id: Long = 0L,
     val description: String? = null,
     val sunLevel: Int,
     val moistureLevel: Int,
     val diagnosedAt: LocalDateTime,
-    val plantId: Int,
-    val listOfMedia: List<Int> = emptyList(),
-    val routinesIds: List<Int> = emptyList()
+    val plantId: Long,
+    val listOfMedia: List<Long> = emptyList(),
+    val routinesIds: List<Long> = emptyList()
 )
